fakedata changed ,remove middleware to become one controller and edit sql query
