+++ server/src/database/config/connection.js
@@ -16,7 +16,7 @@ if (!dbUrl) throw new Error('No Database FOUND');
 
 const options = {
   connectionString: dbUrl,
-  ssl: true,
+  ssl: { rejectUnauthorized: false },
 };
 
 module.exports = new Pool(options);
