+++ server/src/database/queries/checkEmail.js
@@ -2,7 +2,7 @@ const connection = require('../config/connection');
 
 const checkEmail = (email) => {
   const sql = {
-    text: 'select * from bookinguser where email = $1',
+    text: 'SELECT * FROM bookinguser WHERE email = $1',
     values: [email],
   };
   return connection.query(sql);
