+++ server/src/utils/jwt.js
@@ -1,0 +1,14 @@
+const jwt = require('jsonwebtoken');
+
+const sendCookies = (payload) =>
+  new Promise((resolve, reject) => {
+    jwt.sign(payload, process.env.SECRET_KEY, (error, token) => {
+      if (error) {
+        reject(error);
+      } else {
+        resolve(token);
+      }
+    });
+  });
+
+module.exports = { sendCookies };
